### 2025-09-23 #9067 @editorialbot
[PRE REVIEW]: folie: Finding Optimal Langevin Inferred Equations
**Submitting author:** @HadrienNU (<a href="http://orcid.org/0000-0002-2443-5901">Hadrien Vroylandt</a>)
**Repository:** https://github.com/langevinmodel/folie
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0
