### 2025-09-24 a4d3b49
Update paper.md
