### 2025-09-24 7cda109
Update doi in bib
