+++ paper/paper.bib
@@ -62,18 +62,30 @@ bookTitle="The Fokker-Planck Equation: Methods of Solution and Applications",
 year="1996",
 publisher="Springer Berlin Heidelberg",
 address="Berlin, Heidelberg",
-doi="https://doi.org/10.1007/978-3-642-61544-3",
+doi="10.1007/978-3-642-61544-3",
 }
 
 @article{hoffmann2021deeptime,
-  title={Deeptime: a Python library for machine learning dynamical models from time series data},
-  author={Hoffmann, Moritz and Scherer, Martin Konrad and Hempel, Tim and Mardt, Andreas and de Silva, Brian and Husic, Brooke Elena and Klus, Stefan and Wu, Hao and Kutz, J Nathan and Brunton, Steven and Noé, Frank},
-  journal={Machine Learning: Science and Technology},
-  year={2021},
-  publisher={IOP Publishing}
+doi = {10.1088/2632-2153/ac3de0},
+url = {https://dx.doi.org/10.1088/2632-2153/ac3de0},
+year = {2021},
+month = {dec},
+publisher = {IOP Publishing},
+volume = {3},
+number = {1},
+pages = {015009},
+author = {Hoffmann, Moritz and Scherer, Martin and Hempel, Tim and Mardt, Andreas and de Silva, Brian and Husic, Brooke E and Klus, Stefan and Wu, Hao and Kutz, Nathan and Brunton, Steven L and Noé, Frank},
+title = {Deeptime: a Python library for machine learning dynamical models from time series data},
+journal = {Machine Learning: Science and Technology},
 }
 
-@article{kirkby2024pymle, title={pymle: A Python Package for Maximum Likelihood Estimation and Simulation of Stochastic Differential Equations}, author={Kirkby, Justin and Nguyen, Dang and Nguyen, Duy and Nguyen, Nhu N}, journal={Journal of Statistical Software, Forthcoming}, year={2025} }
+@article{kirkby2024pymle, 
+title={pymle: A Python Package for Maximum Likelihood Estimation and Simulation of Stochastic Differential Equations}, 
+author={Kirkby, Justin and Nguyen, Dang and Nguyen, Duy and Nguyen, Nhu N}, 
+journal = {Journal of Statistical Software}, 
+year = {2025}, volume = {113}, number = {4}, pages = {1--40}, 
+doi ={10.18637/jss.v113.i04}
+}
 
 @article{mouaffac2023,
   title = {Optimal {{Reaction Coordinates}} and {{Kinetic Rates}} from the {{Projected Dynamics}} of {{Transition Paths}}},
