### 2025-09-24 6f21b9c
remove github link from bib
