+++ paper/paper.bib
@@ -1,12 +1,3 @@
-
-@misc{GitHub_optLE,
-  howpublished = "\url{https://github.com/physix-repo/optLE}"
-}
-
-@misc{GitHub_colvars,
-  howpublished = "\url{https://github.com/Colvars/colvars}"
-}
-
 @article{PalacioRodriguez2022,
 author = {Palacio-Rodriguez, Karen and Pietrucci, Fabio},
 title = {Free Energy Landscapes, Diffusion Coefficients, and Kinetic Rates from Transition Paths},
