+++ folie/estimation/overdamped_transitionDensity.py
@@ -84,7 +84,7 @@ class EulerDensity(TransitionDensity):
 
         # TODO: Add correction terms
         E = x + self._model.drift(x, bias, **kwargs) * dt
-        V = 2 * (self._model.diffusion(x, **kwargs)) * dt
+        V = 2 * (self._model.diffusion(x[:, : self._model.dim_x], **kwargs)) * dt 
         if not self.use_jac:
             ll = gaussian_likelihood_ND(xt, E, V)
             return ll, np.zeros(2)
@@ -118,22 +118,22 @@ class EulerDensity(TransitionDensity):
         EVE = np.einsum("tdh,tdf,tfg-> thg", E2, invV, E2)
         EV = np.einsum("tdh,tdf-> thf", E2, invV)
         VE = np.einsum("tdf,tfh-> tdh", invV, E2)
-
-        extra_ll = -0.5 * np.einsum("tij,tji->t", invV[:, dh:, dh:], dhdh) + 0.5 * np.einsum("tij,tji->t", EV[:, :, dh:], dhh) + 0.5 * np.einsum("tij,tji->t", VE[:, dh:, :], hdh) - 0.5 * np.einsum("tij,tji->t", EVE, hh)
+        dx = self._model.dim_x
+        extra_ll = -0.5 * np.einsum("tij,tji->t", invV[:, dx:, dx:], dhdh) + 0.5 * np.einsum("tij,tji->t", EV[:, :, dx:], dhh) + 0.5 * np.einsum("tij,tji->t", VE[:, dx:, :], hdh) - 0.5 * np.einsum("tij,tji->t", EVE, hh)
         jacE2 = self._model.friction.grad_coeffs(x[:, : self._model.dim_x], **kwargs) * dt
 
         EVjE = np.einsum("tdh,tdf,tfgc-> thgc", E2, invV, jacE2)
         jEV = np.einsum("tdhc,tdf-> thfc", jacE2, invV)
         VjE = np.einsum("tdf,tfhc-> tdhc", invV, jacE2)
 
-        l_jac_E = -0.5 * np.einsum("tijc,tji->tc", jEV[:, :, dh:, :], dhh) - 0.5 * np.einsum("tijc,tji->tc", VjE[:, dh:, ...], hdh) + np.einsum("tijc,tji->tc", EVjE, hh)
+        l_jac_E = -0.5 * np.einsum("tijc,tji->tc", jEV[:, :, dx:, :], dhh) - 0.5 * np.einsum("tijc,tji->tc", VjE[:, dx:, ...], hdh) + np.einsum("tijc,tji->tc", EVjE, hh)
 
         jacV = 2 * self._model.diffusion.grad_coeffs(x[:, : self._model.dim_x], **kwargs) * dt
         jacinvV = -np.einsum("tij,tjkc,tkl->tilc", invV, jacV, invV)
         EjVE = np.einsum("tdh,tdfc,tfg-> thgc", E2, jacinvV, E2)
         EjV = np.einsum("tdh,tdfc-> thfc", E2, jacinvV)
         jVE = np.einsum("tdfc,tfh-> tdhc", jacinvV, E2)
-        l_jac_V = 0.5 * np.einsum("tijc,tji->tc", jacinvV[:, dh:, dh:, :], dhdh) - 0.5 * np.einsum("tijc,tji->tc", jVE[:, dh:, :], hdh) - 0.5 * np.einsum("tijc,tji->tc", EjV[:, :, dh:], dhh) + 0.5 * np.einsum("tijc,tji->tc", EjVE, hh)
[…]
