### 2025-11-06 7637e53
Reduce word count
