+++ paper/paper.md
@@ -98,48 +98,35 @@ The precise form of the transition density is contingent on the choice of a spec
 
 ## Features
 
-The library has been implemented with a modular structure, so that users can assemble the relevant components in simple Python scripts tailored to their needs.
+The library is built with a modular structure, allowing users to assemble components into Python scripts suited to their needs.
 
 - **Model of Overdamped Langevin Dynamics**
-    Several models of possible implementation of Overdamped Langevin equations are present starting from the parent python class $\texttt{Overdamped}$, followed by the implementation of particular cases such as $\texttt{BrownianMotion}$ and $\texttt{OrnsteinUhlenbeck}$. The underdamped case is still under development.
+    Several models of possible implementation of Overdamped Langevin equations are present, followed by the implementation of particular cases ($\texttt{BrownianMotion}$, $\texttt{OrnsteinUhlenbeck}$) , O. The underdamped case is still under development.
 
-- **Model for the force and diffusion coefficient functions**
-    In addition to the dynamics, building the transition density and likelihood estimator requires the specification of the drift and space-dependent diffusion coefficient. To that effect, the $\texttt{Function}$ class has been implemented, which offers functional forms such as polynomials, splines, etc.
+- **Force and diffusion coefficient functions**
+    Completely defining the dynamical model requires the specification of the drift and space-dependent diffusion coefficient. To that effect, the $\texttt{Function}$ class has been implemented, which offers functional forms such as polynomials, splines, etc.
 
 - **Transition densities**
-    Different methods aimed at approximating the form of the propagator are implemented[@iacus2008a]. They require as input a $\texttt{Model}$ object whose drift and diffusion will be used to compute the mean, variance, and, in case of the Elerian transition probability density, the additional parameters to obtain the relative likelihood.
-    These probability densities are later fed to the $\texttt{Likelihood}$ estimator object, which optimizes the drift and diffusion parameters.
+    Several approximation methods for the propagator are implemented[@iacus2008a].  These probability densities are later fed to the $\texttt{Likelihood}$ estimator object, which optimizes the drift and diffusion parameters.
 
  - **Estimation**
-    Given as input the probability densities used to compute the likelihood of the observed trajectories, an estimator object is created.
-    The class of estimator objects playing a central role in performing the addressed task is the $\texttt{LikelihoodEstimator}$ class, within it, the MLE estimators are recovered by making use of the [optimize.minimize()](https://docs.scipy.org/doc/scipy-1.13.1/reference/generated/scipy.optimize.minimize.html) method from the scipy library applied to the negative of the log-likelihood function eq.(\ref{eq:log_likelihood_general}).
+    The $\texttt{LikelihoodEstimator}$ class performs parameter estimation by maximizing the likelihood of observed trajectories, build from transition density, using [scipy.optimize.minimize()](https://docs.scipy.org/doc/scipy-1.13.1/reference/generated/scipy.optimize.minimize.html) on the negative log-likelihood function eq (\ref{eq:log_likelihood_general}).
 
 
 - **Simulation**
-    In addition to the principal purpose of training the maximum-likelihood estimator for a given set of input trajectories, the FOLIE module also allows to simulate trajectories, which is useful for creating synthetic data to be used when developing methods. This is achieved by first specifying the model of the Langevin equation guiding the evolution of the system through a suitable $\texttt{Overdamped}$ object. Then this passed to the $\texttt{Simulator}$ (or possibly $\texttt{BiasedSimulator}$) class specifying the integration timestep employed. Finally, the dataset is generated by calling $\texttt{Simulator.run()}$.
-
[…]
