### 2025-11-28 6558c7b
Further shortening of the text
