+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'folie: Finding Optimal Langevin Inferred Equations'
+title: 'FOLIE: Finding Optimal Langevin Inferred Equations'
 tags:
   - Python
   - molecular dynamics
@@ -48,7 +48,8 @@ Such low-dimensional models are constructed based on a projection of the full dy
 FOLIE is designed to allow easy and efficient inference of such models from projected molecular simulations. There exits several software packages performing related tasks, but FOLIE differ mainly by its flexibility in the description of the energy landscape and its modular construction for the estimation task.
 [DeepTime](https://deeptime-ml.github.io) (previously pyEmma) [@hoffmann2021deeptime] is an equivalent for discrete Markovian processes. [pymle](https://github.com/jkirkby3/pymle) fits continuous SDEs but imposes restraints on the underlying energy landscapes, which are better suited to econometrics and financial markets [@kirkby2024pymle].
 [OptLE](https://github.com/physix-repo/optle) [@PalacioRodriguez2022] is an experimental Fortran package that inspired this work. It focuses on method development and was not designed for flexibility or scalability.
-[pyOptLE](https://github.com/jhenin/pyOptLE) was a first attempt at improving scalability, with a very limited scope. [StochasticForceInference](https://github.com/ronceray/StochasticForceInference) and [UnderdampedLangevinInference](https://github.com/ronceray/UnderdampedLangevinInference) are codes from the same group focusing on the similar task of fitting continuous SDE but focusing on biological applications and such having less flexibility on the underlying energy landscapes.
+[pyOptLE](https://github.com/jhenin/pyOptLE) was our first attempt at improving scalability, with a limited scope.
+[StochasticForceInference](https://github.com/ronceray/StochasticForceInference) and [UnderdampedLangevinInference](https://github.com/ronceray/UnderdampedLangevinInference) from the Ronceray group perform the task of fitting continuous SDE, with a focus on biological applications and less flexibility in the underlying energy landscapes.
 
 
 # Theoretical background
@@ -56,77 +57,78 @@ FOLIE is designed to allow easy and efficient inference of such models from proj
 
 ## Langevin Models
 
-Let's consider the dynamics of a low-dimensional collective variable $q$. There is a set of possible Langevin models to describe this dynamics[@PalacioRodriguez2022,@girardier2023]. Projecting the high-dimensionnal dynamics onto the collective variable leads to the generalized Langevin equation[@vroylandt2022a] that features a memory kernel representing the complex interactions with the environment. Assuming that the timescale of evolution of the collective variable is slow with respect to its environment. We then obtain the memory-less (Markovian) Standard Langevin equation, that is a second order stochastic differential equation. If one were to assume the inertial time scale is small, the inertial effect are quickly damped. This further approximation leads to the Overdamped Langevin Equation
+Several types of Langevin equations may be relevant for describing projected dynamics[@PalacioRodriguez2022,@girardier2023].
+Projecting high-dimensional dynamics onto a collective variable $q$ leads to the generalized Langevin equation[@vroylandt2022a] that features a memory kernel. Assuming that the timescale of evolution of $q$ is slow with respect to its environment gives the memory-less (Markovian) Standard Langevin equation. If the inertial effect are quickly damped, a further approximation can be made, leading to the Overdamped Langevin Equation
 \begin{equation}
 \dot{q}= -\beta D(q)\frac{\partial A(q)}{\partial q}+ \frac{\partial D(q)}{\partial q} + \sqrt{2D(q)}\eta(t)
 \end{equation}
[…]
