### 2025-12-12 #9067 @mbobra
Thank you for agreeing to review, @joaovaleriano ! You can find reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), together with all the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html). Here are the next steps: I will find a second reviewer and then I will begin the review in January, after the holidays.
