### 2026-01-12 #9748 @editorialbot
[REVIEW]: folie: Finding Optimal Langevin Inferred Equations
**Submitting author:** @HadrienNU (<a href="http://orcid.org/0000-0002-2443-5901">Hadrien Vroylandt</a>)
**Repository:** https://github.com/langevinmodel/folie
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0
