### 2026-01-20 36e8e95
Update notebooks after refactor
Also make free_energy_profile_1d robust to ordering of x points.
