### 2026-02-02 #9748 @joaovaleriano
Hi! I just put a first review at [lanveginmodel/folie#18](https://github.com/langevinmodel/folie/issues/18#issue-3886061902). In a short summary, I believe FOLIE in its current state could be a general-purpose package for inference of 1D SDEs, maybe with a focus on the Kramers-Moyal estimator, which seems to be the most robust option when applicable. That means I do not see a clear match between the summary and statement of need sections of the paper and the actual repository. Its implementation is quite general and does not seem particularly targeted at the community studying molecular dynamics via collective variables. Also, it does not seem quite straightforward how to apply it to >1D systems, which seems like an important limitation. I think the authors should decide if they prefer to rebrand FOLIE according to its current implementation, or if they would like to change the package to match its current purpose, as in the paper. Or if I misunderstood something, then please correct me, and I would be happy to look at it.

Regardless of the match between purpose and implementation, I believe the package could benefit from some effort on documentation, in particularly adding docstrings specifying the meaning and type of attributes. I also think it would be important to add at least one example that is very detailed and commented, with clear explanations of outputs, and this would be particularly useful if it were an example of how to work with custom spatial dependence functions, which is something I believe users would appreciate, and would make a good tutorial for the understanding of how the different modules of FOLIE connect, making it easier for the community to contribute to the project.

More details in the mentioned issue. Happy to discuss further!
