### 2026-03-04 745a948
Construct meshed domains for 2D systems
