+++ folie/domains/distmesh.py
@@ -1,0 +1,249 @@
+"""
+DistMesh Adapted from distmesh librairy from Per-Olof Persson and  Bradley Froehle
+using Gemini to simplify the code
+
+"""
+
+import numpy as np
+import scipy.spatial as spspatial
+from scipy.spatial.distance import cdist
+import itertools
+import matplotlib.pyplot as plt
+from matplotlib.collections import PolyCollection
+from scipy.ndimage import gaussian_filter
+from scipy.interpolate import RegularGridInterpolator
+
+# --- Distance Function Utilities ---
+
+
+def huniform(p):
+    """Implements the trivial uniform mesh size function h=1."""
+    return np.ones(p.shape[0])
+
+
+class hdensity:
+    """
+    Implement density based
+    """
+
+    def __init__(self, X, h0, bbox, min_scale=0.2, alpha=0.5):
+        """
+        Efficient sizing based on density rescaling.
+
+        Parameters:
+        -----------
+        h0 : Background triangle size.
+        min_scale : Finest triangle size relative to h0 (e.g., 0.2 * h0).
+        alpha : Contrast control. 1.0 = Linear, <1.0 = Faster transition.
+        """
+        self.h0 = h0
+        self.min_scale = min_scale
+
+        # 1. Setup Grid Resolution tied to h0
+        xmin, ymin, xmax, ymax = bbox
+        # Padding by 2*h0 to handle boundary interpolation safely
+        self.x_coords = np.arange(xmin - 2 * h0, xmax + 2 * h0, h0)
+        self.y_coords = np.arange(ymin - 2 * h0, ymax + 2 * h0, h0)
+
+        # 2. Compute Density (Histogram)
+        hist, _, _ = np.histogram2d(X[:, 0], X[:, 1], bins=[self.x_coords, self.y_coords])
+
+        # 3. Smooth (Sigma=1.0 spreads density over ~3 cells/h0)
+        smoothed_hist = gaussian_filter(hist, sigma=1.0)
+
+        # 4. LINEAR RESCALING (No Clipping)
+        # We map Density=0 to 1.0 and Density=Max to min_scale
+        d_max = smoothed_hist.max()
+        if d_max > 0:
+            # Normalized density [0, 1]
+            norm_d = (smoothed_hist / d_max) ** alpha
+            # Linear map: h = 1.0 - (1.0 - min_scale) * norm_d
+            sizing_grid = 1.0 - (1.0 - min_scale) * norm_d
+        else:
+            sizing_grid = np.ones_like(smoothed_hist)
+
+        # 5. Fast Interpolator
+        xc = (self.x_coords[:-1] + self.x_coords[1:]) / 2
+        yc = (self.y_coords[:-1] + self.y_coords[1:]) / 2
+        self.interp = RegularGridInterpolator((xc, yc), sizing_grid, method="linear", bounds_error=False, fill_value=1.0)
+
+    def __call__(self, p):
+        # This is the fh(p) function for DistMesh
+        return self.interp(p)
+
+
+def ddiff(d1, d2):
+    return np.maximum(d1, -d2)
+
+
+def dcircle(p, xc, yc, r):
+    return np.sqrt((p[:, 0] - xc) ** 2 + (p[:, 1] - yc) ** 2) - r
+
+
+def dellipse(p, xc, yc, rx, ry):
+    return np.sqrt(((p[:, 0] - xc) ** 2) / rx**2 + ((p[:, 1] - yc) ** 2) / ry**2) - 1
+
+
+def drectangle(p, x1, x2, y1, y2):
[…]

[…]
