+++ folie/analysis/fem.py
@@ -1,0 +1,265 @@
+"""
+Set of analysis methods using Finite Element Method to solve various equation
+"""
+
+import numpy as np
+import skfem
+from skfem import BilinearForm
+from skfem.helpers import grad, dot, mul, inv
+from scipy.sparse.linalg import eigs, eigsh
+
+from ..models import BaseModelOverdamped
+
+
+def free_energy_profile(model, scalar_basis, x=None, v_basis=None):
+    r"""
+    Construct the free energy profile V(x) from model force F(x) and diffusion D(x).
+    Formula: F(x) = -D(x) grad V(x) + div D(x)
+    => grad V(x) = D(x)^-1 * (div D(x) - F(x))
+    """
+    if v_basis is None:
+        v_basis = scalar_basis
+
+    @skfem.LinearForm
+    def b_term(v, w):
+        X = w["x"].reshape(w["x"].shape[0], -1).T
+        F = model.pos_drift(X).T.reshape(w["x"].shape)
+        D = model.diffusion(X).T.reshape((w["x"].shape[0], w["x"].shape[0], *w["x"].shape[1:]))
+        if D.shape[0] == 1:  # 1D case that is not taken into account by skfem
+            D_inv = 1.0 / D
+        else:
+            D_inv = inv(D)
+
+        grad_D = model.diffusion.grad_x(X).T.reshape((w["x"].shape[0], w["x"].shape[0], w["x"].shape[0], *w["x"].shape[1:]))
+        div_D = np.einsum("jij...->i...", grad_D)
+
+        # RHS vector field: R = D^-1 * (div_D - F)
+        R = mul(D_inv, div_D - F)
+
+        if isinstance(v_basis.elem, skfem.ElementVector):
+            return dot(v, R)
+        else:
+            return dot(v.grad, R)
+
+    if isinstance(v_basis.elem, skfem.ElementVector):
+        basis_product = BilinearForm(lambda u, v, w: dot(u.grad, v))
+    else:
+        basis_product = BilinearForm(lambda u, v, w: dot(u.grad, v.grad))
+
+    A = basis_product.assemble(scalar_basis, v_basis)
+    b = b_term.assemble(v_basis)
+
+    # Handle singularity: fix the first DOF to 0 to determine the constant shift
+    U_coeff = skfem.solve(*skfem.condense(A, b, D=np.array([0])))
+
+    if x is not None:
+        if isinstance(x, str) and x == "nodal":
+            return U_coeff[scalar_basis.nodal_dofs]
+        return scalar_basis.probes(x.T) @ U_coeff
+    return U_coeff
+
+
+# def free_energy_profile(model, scalar_basis, x=None, v_basis=None):
+#     r"""
+#     From the model force F(x) and diffusion D(x) construct the free energy profile V(x) using the formula
+
+#     .. math::
+#         F(x) = -D(x) \nabla V(x) + \mathrm{div} D(x)
+
+#     Parameters
+#     ----------
+#     scalar_basis: skfem basis
+#         Should be a scalar basis with the same quadrature point than internal basis
+#     x: ndarray, default None
+#         Evaluation points. If x is None, evaluate at nodals points of the basis
+#     """
+#     if v_basis is None:
+#         v_basis = scalar_basis
+#     if isinstance(v_basis, skfem.ElementVector):
+
+#         @skfem.LinearForm
+#         def b_term(v, w):
+#             X = w["x"].reshape(w["x"].shape[0], -1).T
+#             F = model.pos_drift(X).T.reshape(*w["x"].shape)
+#             D = model.diffusion(X).T.reshape(w["x"].shape[0], w["x"].shape[0], *w["x"].shape[1:])
[…]
