+++ folie/analysis/overdamped_1d.py
@@ -27,12 +27,7 @@ def free_energy_profile_1d(model, x):
     idx = np.argsort(x)
     x_sorted = x[idx]
 
-    sol = solve_ivp(
-        grad_V,
-        [x_sorted.min() - 1e-10, x_sorted.max() + 1e-10],
-        np.array([0.0]),
-        t_eval=x_sorted
-    )
+    sol = solve_ivp(grad_V, [x_sorted.min() - 1e-10, x_sorted.max() + 1e-10], np.array([0.0]), t_eval=x_sorted)
 
     # Re-map the potential V back to the original order of x
     V_sorted = sol.y.ravel()
@@ -102,14 +97,17 @@ def mfpt_1d(model, x_end: float, x_range, Npoints=500, x_start=None):
         # Compute lower part
         int_range = np.linspace(x_range[0], x_end, Npoints)
         prob_well = cumulative_trapezoid(np.exp(-free_energy_profile_1d(model, int_range)), int_range, initial=0.0)
-
         x_int_lower = np.linspace(x_end, x_range[0], Npoints)
-        res_lower = -1 * cumulative_trapezoid(np.exp(free_energy_profile_1d(model, x_int_lower)) * np.interp(x_int_lower, int_range, prob_well) / model.diffusion(x_int_lower.reshape(-1, 1)).ravel(), x_int_lower, initial=0.0)
+        res_lower = -1 * cumulative_trapezoid(
+            np.exp(free_energy_profile_1d(model, x_int_lower))[::-1] * np.interp(x_int_lower, int_range, prob_well) / model.diffusion(x_int_lower.reshape(-1, 1)).ravel(), x_int_lower, initial=0.0
+        )
 
         int_range = np.linspace(x_end, x_range[1], Npoints)
         prob_well = -1 * cumulative_trapezoid(np.exp(-free_energy_profile_1d(model, int_range)), int_range, initial=0.0)
         prob_well -= prob_well[-1]
-        # return (int_range, prob_well)
         x_int_upper = np.linspace(x_end, x_range[1], Npoints)
-        res_upper = cumulative_trapezoid(np.exp(free_energy_profile_1d(model, x_int_upper)) * np.interp(x_int_upper, int_range, prob_well) / model.diffusion(x_int_upper.reshape(-1, 1)).ravel(), x_int_upper, initial=0.0)
+        res_upper = cumulative_trapezoid(
+            np.exp(free_energy_profile_1d(model, x_int_upper)) * np.interp(x_int_upper, int_range, prob_well) / model.diffusion(x_int_upper.reshape(-1, 1)).ravel(), x_int_upper, initial=0.0
+        )
+
         return np.hstack((x_int_lower[::-1], x_int_upper)), np.hstack((res_lower[::-1], res_upper))
