### 2026-05-12 703a9c0
Add 2d FEM example and pcca++ analysis
