+++ examples/plot_fem_2d_committor.py
@@ -1,0 +1,178 @@
+"""
+========================================
+2D Muller-Brown — FEM Committor & PCCA
+========================================
+
+Compute committor probabilities and reduced Markov state kinetics
+for a 2D Muller-Brown potential using FEM eigenmode decomposition and PCCA++.
+
+Demonstrates:
+- solve_committor_fem() for committor between two basins
+- reduced_matrix() for PCCA++ coarse-graining
+- Eigenmode visualization for collective variables
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+import folie as fl
+import skfem
+from scipy.sparse.linalg import eigs
+
+# ---- Define 2D Muller-Brown model ----
+
+# Minima at (~1, 0), (~0, 0.5), (~-0.5, 1.5)
+D = 1.0
+
+muller_brown = fl.functions.MullerBrown()
+diff_function = fl.functions.Polynomial(deg=0, coefficients=D * np.eye(2, 2), output_shape=(2, 2))
+
+model = fl.models.overdamped.Overdamped(muller_brown, diffusion=diff_function)
+
+# ---- Create FEM mesh and build matrices ----
+
+# Use a mesh that covers the Muller-Brown minima region
+x_min, x_max = -1.5, 1.5
+y_min, y_max = -0.5, 2.0
+n_x, n_y = 25, 25
+
+x_range = np.linspace(x_min, x_max, n_x)
+y_range = np.linspace(y_min, y_max, n_y)
+
+mesh = skfem.MeshTri.init_tensor(x_range, y_range)
+# ---- 4. Define subdomains (reactant and product basins) ----
+
+# Reactant basin around global minimum (-0.5, 0), product basin around local minimum (0.5, 0)
+mesh = mesh.with_subdomains(
+    {
+        "reactant": lambda X: (X[0] + 0.5) ** 2 + (X[1] - 1.5) ** 2 < 0.05,
+        "product": lambda X: (X[0] - 0.5) ** 2 + X[1] ** 2 < 0.05,
+    }
+)
+
+
+# Build FEM matrices
+A, M, basis = fl.analysis.fem.build_fem_matrices(model, mesh, element=skfem.ElementTriP1())
+print(f"FEM mesh: {mesh.p.shape[1]} nodes, {mesh.t.shape[1]} elements")
+print(f"FEM basis: {basis.N} DOFs")
+
+reactant_dofs = basis.get_dofs(elements="reactant")
+product_dofs = basis.get_dofs(elements="product")
+print(f"Reactant DOFs: {len(reactant_dofs)}, Product DOFs: {len(product_dofs)} (element subdomains, not facets)")
+
+# ---- Solve committor equation ----
+
+u_sol, u_bc, fem_basis = fl.analysis.fem.solve_committor_fem(model, mesh, bc="elements")
+
+# Evaluate committor on grid
+n_plot = 150
+x_plot = np.linspace(x_min, x_max, n_plot)
+y_plot = np.linspace(y_min, y_max, n_plot)
+X_grid, Y_grid = np.meshgrid(x_plot, y_plot)
+X_eval = np.column_stack([X_grid.ravel(), Y_grid.ravel()])
+
+committor_nodal = u_sol[fem_basis.nodal_dofs]
+# For tensor meshes, nodal_dofs may be empty; always use probes for 2D
+committor_grid = (fem_basis.probes(X_eval.T) @ u_sol).reshape(n_plot, n_plot)
+
[…]
