### 2026-05-12 4ba6f2d
Documentation of functions
