+++ folie/simulations/stepper.py
@@ -95,7 +95,7 @@ class VECStepper(Stepper):
             gamma = self.model.friction(x)
             diff = self.model.diffusion(x)
 
-        sc2 = 1 - 0.5 * gamma * dt + 0.125 * gamma * dt ** 2
+        sc2 = 1 - 0.5 * gamma * dt + 0.125 * gamma * dt**2
         c1 = 0.5 * dt * (1 - 0.25 * gamma * dt)
         d1 = 0.5 * (1 - 0.25 * gamma * dt)
         d2 = -0.25 * gamma * dt / np.sqrt(3)
@@ -111,8 +111,9 @@ class VECStepper(Stepper):
 
         v = ((sc2 * v_mid.T) + c1 * self.f + (d1 * dWx.T) + (d2 * dWv.T)).T
         return np.concatenate([x, v], axis=1)
-    
+
     def run_step_ND(self, X, dt, dW, bias=0.0):
+        print(self.model.dim)
         x = X[:, : self.model.dim]
         v = X[:, self.model.dim :]
         # If not initialized initialize
@@ -124,24 +125,23 @@ class VECStepper(Stepper):
             fx = self.model.pos_drift(x, bias)
             gamma = self.model.friction(x)
             diff = self.model.diffusion(x)
-            
 
         N, dim = gamma.shape[0], self.model.dim
-        sc2 = np.broadcast_to(np.eye(dim), (N, dim, dim)) - 0.5 * gamma * dt + 0.125 * gamma * dt ** 2 # prev. version : (gamma * dt) ** 2 does not match VEC paper
+        sc2 = np.broadcast_to(np.eye(dim), (N, dim, dim)) - 0.5 * gamma * dt + 0.125 * gamma * dt**2  # prev. version : (gamma * dt) ** 2 does not match VEC paper
         c1 = 0.5 * dt * (np.broadcast_to(np.eye(dim), (N, dim, dim)) - 0.25 * gamma * dt)
         d1 = 0.5 * (np.broadcast_to(np.eye(dim), (N, dim, dim)) - 0.25 * gamma * dt)
         d2 = -0.25 * gamma * dt / np.sqrt(3)
         sig_sq_dt = np.sqrt(2 * diff * dt)
-        dWx = np.einsum('ijk,ik->ij', sig_sq_dt, dW[:, : self.model.dim])
-        dWv = np.einsum('ijk,ik->ij', sig_sq_dt, dW[:, self.model.dim :])
-        v_mid = np.einsum('ijk,ik->ij', sc2, v) + np.einsum('ijk,ik->ij', c1, fx) + np.einsum('ijk,ik->ij',d1, dWx) + np.einsum('ijk,ik->ij', d2, dWv)
+        dWx = np.einsum("ijk,ik->ij", sig_sq_dt, dW[:, : self.model.dim])
+        dWv = np.einsum("ijk,ik->ij", sig_sq_dt, dW[:, self.model.dim :])
+        v_mid = np.einsum("ijk,ik->ij", sc2, v) + np.einsum("ijk,ik->ij", c1, fx) + np.einsum("ijk,ik->ij", d1, dWx) + np.einsum("ijk,ik->ij", d2, dWv)
         x += dt * (v_mid + (0.5 / np.sqrt(3)) * dWv)
         # update friction
         self.f = self.model.pos_drift(x, bias)
         self.gamma = self.model.friction(x)
         self.diff = self.model.diffusion(x)
 
[…]
