+++ examples/toy_models/plot_2D_Double_Well.py
@@ -18,7 +18,7 @@ y = np.linspace(-1.8, 1.8, 36)
 input = np.transpose(np.array([x, y]))
 
 D = 0.5
-diff_function = fl.functions.Polynomial(deg=0, coefficients=D * np.eye(2, 2))
+diff_function = fl.functions.Polynomial(deg=0, coefficients=D * np.eye(2, 2), output_shape=(2, 2))
 a, b = 5, 10
 drift_quartic2d = fl.functions.Quartic2D(a=D * a, b=D * b)  # simple way to multiply D*Potential here force is the SDE force (meandispl)  ## use this when you need the drift ###
 quartic2d = fl.functions.Quartic2D(a=a, b=b)  # Real potential , here force is just -grad pot ## use this when you need the potential energy ###

[…]
